fix parquetmaker.nf

+++ modules/local/parquetmaker.nf
@@ -20,24 +20,25 @@ process PARQUETMAKER {
     def run_name = run_path.getBaseName()
 
     """
-    # Just cleaning up the logic otherise nt.fasta and aa.fasta will not created and the parquet_maker.py will fail
-    if [[ -f MIRA_${run_name}_failed_amended_consensus.fasta && -f MIRA_${run_name}_amended_consensus.fasta ]]; then
-        cat MIRA_${run_name}_amended_consensus.fasta MIRA_${run_name}_failed_amended_consensus.fasta > nt.fasta
-    elif [[ -f MIRA_${run_name}_amended_consensus.fasta ]]; then
-        cat MIRA_${run_name}_amended_consensus.fasta > nt.fasta
-    elif [[ -f MIRA_${run_name}_failed_amended_consensus.fasta ]]; then
-        cat MIRA_${run_name}_failed_amended_consensus.fasta > nt.fasta
-    else
-        touch nt.fasta  # Create an empty nt.fasta if neither file exists
+    ## This logic is very specific to the MIRA pipeline and should not be changed.
+    ## If aa.fast and nt.fast is not being created then something it broken up stream that needs to be fixed
+    if [[ -f  MIRA_${run_name}_failed_amended_consensus.fasta && -f MIRA_${run_name}_amended_consensus.fasta ]]; then
+    cat MIRA_${run_name}_amended_consensus.fasta MIRA_${run_name}_failed_amended_consensus.fasta > nt.fasta
     fi
-    if [[ -f MIRA_${run_name}_failed_amino_acid_consensus.fasta && -f MIRA_${run_name}_amino_acid_consensus.fasta ]]; then
-         MIRA_${run_name}_amino_acid_consensus.fasta MIRA_${run_name}_failed_amino_acid_consensus.fasta > aa.fasta
-    elif [[ -f MIRA_${run_name}_amino_acid_consensus.fasta ]]; then
-        cat MIRA_${run_name}_amino_acid_consensus.fasta > aa.fasta
-    elif [[ -f MIRA_${run_name}_failed_amino_acid_consensus.fasta ]]; then
-        cat MIRA_${run_name}_failed_amino_acid_consensus.fasta > aa.fasta
-    else
-        touch aa.fasta  # Create an empty aa.fasta if neither file exists
+    if [[ ! -f  MIRA_${run_name}_failed_amended_consensus.fasta && -f MIRA_${run_name}_amended_consensus.fasta ]]; then
+    cat MIRA_${run_name}_amended_consensus.fasta > nt.fasta
+    fi
+    if [[ -f  MIRA_${run_name}_failed_amended_consensus.fasta && ! -f MIRA_${run_name}_amended_consensus.fasta ]]; then
+    cat MIRA_${run_name}_failed_amended_consensus.fasta > nt.fasta
+    fi
+    if [[ -f  MIRA_${run_name}_failed_amino_acid_consensus.fasta && -f  MIRA_${run_name}_amino_acid_consensus.fasta ]]; then
+    cat MIRA_${run_name}_amino_acid_consensus.fasta MIRA_${run_name}_amino_acid_consensus.fasta > aa.fasta
+    fi
+    if [[ ! -f  MIRA_${run_name}_failed_amino_acid_consensus.fasta && -f MIRA_${run_name}_amino_acid_consensus.fasta ]]; then
+    cat MIRA_${run_name}_amino_acid_consensus.fasta > aa.fasta
+    fi
+    if [[ -f  MIRA_${run_name}_failed_amino_acid_consensus.fasta && ! -f MIRA_${run_name}_amino_acid_consensus.fasta ]]; then
+    cat MIRA_${run_name}_failed_amino_acid_consensus.fasta > aa.fasta
     fi
 
     python3 ${projectDir}/bin/parquet_maker.py -f nt.fasta -o ${run_name}_amended_consensus -r ${run_name}
